% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{europeananews}
\alias{europeananews}
\title{Tagged news paper articles from Europeana}
\description{
BIO-tagged news articles in different languages extracted from \url{https://github.com/EuropeanaNewspapers/ner-corpora} 
using \code{\link{europeana_read}}
\itemize{
\item{Dutch from the Koninklijke Bibliotheek}
\item{Austrian from the National Library of Austria}
\item{German from the Dr. Friedrich Teßmann Library}
\item{French from the National Library of France, in cooperation with LIP6-ACASA}
}
}
\examples{
data(europeananews)
str(europeananews)
}
\references{
Europeana Newspapers project, (2014), KB Europeana Newspapers NER Dataset. KB Lab: The Hague. \url{https://lab.kb.nl/dataset/europeana-newspapers-ner}
}
