% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nametagger.R
\name{nametagger_load_model}
\alias{nametagger_load_model}
\title{Load a Named Entity Recognition}
\usage{
nametagger_load_model(file)
}
\arguments{
\item{file}{character string with the path to the file on disk}
}
\value{
an object of class nametagger
}
\description{
Load a Named Entity Recognition from your hard disk
}
\examples{
path  <- system.file(package = "nametagger", "models", "exampletagger.ner")
model <- nametagger_load_model(path)
model
}
