% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nametagger.R
\name{write_nametagger}
\alias{write_nametagger}
\title{Save a tokenised dataset as nametagger train data}
\usage{
write_nametagger(x, file = tempfile(fileext = ".txt", pattern = "nametagger_"))
}
\arguments{
\item{x}{a tokenised data.frame with columns doc_id, sentence_id, token containing 1 row per token. In addition
it can have columns lemma and pos representing the lemma and the parts-of-speech tag of the token}

\item{file}{the path to the file where the training data will be saved}
}
\value{
invisibly a character vector of text in the nametagger format
}
\description{
Save a tokenised dataset as nametagger train data
}
\examples{
data(europeananews)
x <- subset(europeananews, doc_id \%in\% "enp_NL.kb.bio")
x <- head(x, n = 250)

bio <- write_nametagger(x, file = "traindata.txt")
str(bio)
}
